Factor out some of the shared parts of nodeLibE2eGenerator
