+++ packages/nx-plugin/src/generators/lib/e2e/index.ts
@@ -1,0 +1,2 @@
+export * from "./libE2eGenerator";
+export * from "./libE2eGenerator.schema";
