+++ packages/nx-plugin/src/generators/lib/e2e/libE2eGenerator.schema.ts
@@ -1,0 +1,8 @@
+export interface LibE2eGeneratorSchema extends LibE2eGeneratorBaseSchema {
+  codeSamplePath: string;
+}
+
+export interface LibE2eGeneratorBaseSchema {
+  name: string;
+  libName: string;
+}
